[REFACTOR]: elements for LoginPage
