+++ e2e-cypress/cypress/e2e/crear-post-4.cy.js
@@ -1,5 +1,5 @@
-import { LoginPage } from './LoginPage';
-import { PostsPage } from './PostPage';
+import { PostsPage } from './pages/PostPage';
+import { LoginPage } from './pages/LoginPage';
 
 describe('Create a post', () => {
     const loginPage = new LoginPage();
